requirements.txt
pandas
numpy
faker
fhir.resources
random
datetime
kafka-python
datetime
json
os
time